Agreement Coefficients for 3 Raters or More
This script file contains R functions for computing various agreement coefficients for 3 raters or more, when the raw ratings are provided for each rater and subject.  The input data is presented in the form of a data file where each column represents a rater, and each row a subject.  Each cell is the rating that the rater assigned to the subject.