Agreement coefficients without raw ratings
This script file contains R functions for computing various agreement coefficients when the input data is in the form of the distribution of raters by subject and by category.  Each column in the dataset represents a category, while each row represents a subject.  Each cell on the other hand contains the number of raters who classified the subject into the category.